+++ src/Facades/ActiveCampaign.php
@@ -3,6 +3,8 @@
 namespace Datomatic\ActiveCampaign\Facades;
 
 use Datomatic\ActiveCampaign\Resources\ActiveCampaignContactsResource;
+use Datomatic\ActiveCampaign\Resources\ActiveCampaignFieldOptionsResource;
+use Datomatic\ActiveCampaign\Resources\ActiveCampaignFieldRelsResource;
 use Datomatic\ActiveCampaign\Resources\ActiveCampaignFieldsResource;
 use Datomatic\ActiveCampaign\Resources\ActiveCampaignFieldValuesResource;
 use Datomatic\ActiveCampaign\Resources\ActiveCampaignTagsResource;
@@ -14,6 +16,8 @@ use Illuminate\Support\Facades\Facade;
  * @method static ActiveCampaignContactsResource contacts()
  * @method static ActiveCampaignFieldValuesResource fieldValues()
  * @method static ActiveCampaignFieldsResource fields()
+ * @method static ActiveCampaignFieldOptionsResource fieldOptions()
+ * @method static ActiveCampaignFieldRelsResource fieldRels()
  * @method static ActiveCampaignTagsResource tags()
  */
 class ActiveCampaign extends Facade